+++ src/Krystal/Application/FrontController/ControllerFactoryInterface.php
@@ -17,8 +17,9 @@ interface ControllerFactoryInterface
      * Builds a controller instance
      * 
      * @param string $controller PSR-0 Compliant path
-     * @param array $options Route options
+     * @param string $action Method to be invoked on controller
+     * @param array $options Route options passed to corresponding controller
      * @return \Krystal\Application\Controller\AbstractController
      */
-    public function build($controller, array $options);
+    public function build($controller, $action, array $options);
 }